added Float() to max_mark to fix type error

+++ app/models/assignment.rb
@@ -267,7 +267,7 @@ class Assignment < Assessment
 
   # Returns the maximum possible mark for a particular assignment
   def max_mark(user_visibility = :ta_visible)
-    criteria.where(user_visibility => true, bonus: false).sum(:max_mark).round(2)
+    Float(criteria.where(user_visibility => true, bonus: false).sum(:max_mark).round(2))
   end
 
   # Returns a boolean indicating whether marking has started for at least
